Renamed dvl, ahrs, backplane, and pressure depth transducer files to fit the naming conventions set by this repo.
